+++ src/drivers/sdl/lv_sdl_window.c
@@ -46,7 +46,7 @@ typedef struct {
     uint8_t * rotated_buf;
     size_t rotated_buf_size;
 #endif
-    uint8_t zoom;
+    float zoom;
     uint8_t ignore_size_chg;
 } lv_sdl_window_t;
 
@@ -144,7 +144,7 @@ void lv_sdl_window_set_resizeable(lv_display_t * disp, bool value)
     SDL_SetWindowResizable(dsc->window, value);
 }
 
-void lv_sdl_window_set_zoom(lv_display_t * disp, uint8_t zoom)
+void lv_sdl_window_set_zoom(lv_display_t * disp, float zoom)
 {
     lv_sdl_window_t * dsc = lv_display_get_driver_data(disp);
     dsc->zoom = zoom;
@@ -152,7 +152,7 @@ void lv_sdl_window_set_zoom(lv_display_t * disp, uint8_t zoom)
     lv_refr_now(disp);
 }
 
-uint8_t lv_sdl_window_get_zoom(lv_display_t * disp)
+float lv_sdl_window_get_zoom(lv_display_t * disp)
 {
     lv_sdl_window_t * dsc = lv_display_get_driver_data(disp);
     return dsc->zoom;
